Started splitting out server files
